requirements.txt
Package                 Version      Editable project location
----------------------- ------------ -------------------------
absl-py                 1.3.0
affine                  2.3.1
asttokens               2.0.8
attrs                   22.1.0
backcall                0.2.0
batchgenerators         0.24
cachetools              5.2.0
certifi                 2022.9.14
cffi                    1.15.1
charset-normalizer      2.1.1
click                   8.1.3
click-plugins           1.1.1
cligj                   0.7.2
configobj               5.0.6
configparser            5.3.0
contourpy               1.0.5
cycler                  0.11.0
debugpy                 1.5.1
decorator               5.1.1
dicom2nifti             2.4.3
entrypoints             0.4
et-xmlfile              1.1.0
executing               1.0.0
fonttools               4.37.2
future                  0.18.2
google-auth             2.13.0
google-auth-oauthlib    0.4.6
graphviz                0.20.1
grpcio                  1.50.0
hiddenlayer             0.3
idna                    3.4
imageio                 2.21.3
imio                    0.2.4
imlib                   0.1.3
importlib-metadata      5.0.0
imutils                 0.5.4
ipykernel               6.15.2
ipython                 8.5.0
jedi                    0.18.1
joblib                  1.1.0
jupyter_client          7.3.5
jupyter-core            4.10.0
kiwisolver              1.4.4
linecache2              1.0.0
luddite                 1.0.2
Markdown                3.4.1
MarkupSafe              2.1.1
matplotlib              3.0.3
matplotlib-inline       0.1.6
MedPy                   0.4.0
micrometa               0.0.11
natsort                 8.2.0
nest-asyncio            1.5.5
networkx                2.8.6
nibabel                 4.0.2
nnunet                  1.7.0        /home/ravi/nnUNet
numpy                   1.23.1
oauthlib                3.2.2
opencv-python           4.6.0.66
openpyxl                3.0.10
packaging               21.3
pandas                  1.4.3
parso                   0.8.3
pexpect                 4.8.0
pickleshare             0.7.5
Pillow                  9.2.0
pip                     22.1.2
prompt-toolkit          3.0.31
protobuf                3.19.6
psutil                  5.9.2
ptyprocess              0.7.0
pure-eval               0.2.2
pyasn1                  0.4.8
pyasn1-modules          0.2.8
pycocotools             2.0.5
pycparser               2.21
pydicom                 2.3.0
Pygments                2.13.0
pynrrd                  0.4.3
pyparsing               3.0.9
pytesseract             0.3.10
python-dateutil         2.8.2
python-gdcm             3.0.17.1
pytz                    2022.2.1
pyvips                  2.2.1
PyWavelets              1.4.1
PyYAML                  6.0
pyzmq                   23.2.0
rasterio                1.3.2
requests                2.28.1
requests-oauthlib       1.3.1
rsa                     4.9
scikit-image            0.19.3
scikit-learn            1.1.2
scipy                   1.9.1
seaborn                 0.12.0
setuptools              63.4.1
SimpleITK               2.2.0
six                     1.16.0
sklearn                 0.0
slurmio                 0.0.6
snuggs                  1.4.7
stack-data              0.5.0
tensorboard             2.10.1
tensorboard-data-server 0.6.1
tensorboard-plugin-wit  1.8.1
threadpoolctl           3.1.0
tifffile                2022.8.12
torch                   1.12.1+cu113
torchaudio              0.12.1+cu113
torchvision             0.13.1+cu113
tornado                 6.2
tqdm                    4.64.0
traceback2              1.4.0
traitlets               5.4.0
typing_extensions       4.3.0
unittest2               1.1.0
urllib3                 1.26.12
wcwidth                 0.2.5
Werkzeug                2.2.2
wheel                   0.37.1
zipp                    3.10.0
